Move command modules into cmd directory.
- Move exercise module to root command module
- Rename exercise.go to sweet.go
- Remove root command configuration from main to sweet.go
