+++ cmd/root/results.go
@@ -1,4 +1,4 @@
-package exercise
+package root
 
 import (
 	"fmt"
