--- default-exercises/hello.go
@@ -1,9 +1,0 @@
-package hello
-
-import "fmt"
-
-const Hello = "Hello"
-
-func SayHello() {
-	fmt.Println(Hello)
-}
