--- default-exercises/hello.js
@@ -1,7 +1,0 @@
-export const sayHello = () => {
-  console.log("Hello!")
-}
-
-export const getHello = () => {
-  return "Hello!"
-}
