--- text.txt
@@ -1,3 +1,0 @@
-hi there
-i am text
-AHAA!
